Preferences vertical slice module
