+++ internal/preferences/max_song_duration_preference.go
@@ -1,0 +1,28 @@
+package preferences
+
+import (
+	"context"
+	"time"
+)
+
+func SetMaxSongDurationPreference(durationInMilliseconds int) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
+	err := SetPreference(ctx, MaxDurationKey, durationInMilliseconds)
+	cancel()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func GetMaxSongDurationPreference() (*Preference, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
+	pref, err := GetPreference(ctx, MaxDurationKey)
+	cancel()
+	if err != nil {
+		return nil, err
+	}
+
+	return pref, nil
+}
